feat(tasks): prepare a provider seam for hogland sandboxes (rebase attempt 2)
