+++ products/tasks/backend/logic/services/sandbox.py
@@ -767,6 +767,12 @@ def _get_modal_evals_sandbox_class() -> SandboxClass:
     return ModalEvalsSandbox
 
 
+def _get_hogland_sandbox_class() -> SandboxClass:
+    from .hogland_sandbox import HoglandSandbox
+
+    return HoglandSandbox
+
+
 def get_sandbox_class() -> SandboxClass:
     provider = getattr(settings, "SANDBOX_PROVIDER", None)
 
@@ -778,6 +784,16 @@ def get_sandbox_class() -> SandboxClass:
 
     if provider and provider.upper() == "MODAL_EVALS":
         return _get_modal_evals_sandbox_class()
+
+    if provider and provider.lower() == "hogland":
+        # Global default only for local development — production routes per run via
+        # get_sandbox_class_for_backend, driven by the tasks-hogland-sandbox flag.
+        if not (settings.DEBUG or settings.TEST):
+            raise RuntimeError(
+                "SANDBOX_PROVIDER=hogland is for local development only. In production the "
+                "hogland backend is selected per run by the tasks-hogland-sandbox feature flag."
+            )
+        return _get_hogland_sandbox_class()
 
     # Default to Modal everywhere
     from .modal_sandbox import ModalSandbox
@@ -796,7 +812,34 @@ def get_sandbox_class_for_backend(backend: str) -> SandboxClass:
         return _get_modal_evals_sandbox_class()
     if backend == "docker":
         return _get_docker_sandbox_class()
+    if backend == "hogland":
+        return _get_hogland_sandbox_class()
     raise RuntimeError(f"Unsupported sandbox backend: {backend}")
+
+
+# hogland mints `box-<12 hex>` (hogd enforces `^box-[0-9a-f]{12}$`); Modal object ids
+# are `sb-...`. A box restored from a pen keeps a `box-` id, so this covers pens too.
+HOGLAND_SANDBOX_ID_PREFIX = "box-"
+
+
+def get_sandbox_class_for_sandbox_id(sandbox_id: str) -> SandboxClass:
+    """Resolve the provider class for an existing sandbox from its id alone.
+
+    Hogland box ids are `box-...` and Modal object ids `sb-...`, so the prefix is enough
+    to route the ~20 `get_by_id` call sites that hold only a persisted sandbox id (the
+    reaper, cleanup, and snapshot activities have no other backend context). Anything
+    that is not a hogland id falls through to the process-wide provider, preserving the
+    docker/local-dev behavior.
+
+    Getting this prefix wrong fails closed to the wrong provider: a hogland id would
+    resolve to Modal, whose `get_by_id` raises `SandboxNotFoundError`, so cleanup and the
+    reaper would silently skip a real hogbox and leak it. The persisted `sandbox_backend`
+    (see get_task_processing_context) is the authoritative signal for behavioral branches;
+    this prefix is a routing convenience checked against hogland's enforced id shape.
+    """
+    if sandbox_id.startswith(HOGLAND_SANDBOX_ID_PREFIX):
+        return _get_hogland_sandbox_class()
+    return get_sandbox_class()
 
 
 if TYPE_CHECKING:
