+++ products/tasks/backend/temporal/create_snapshot/activities/cleanup_sandbox.py
@@ -5,7 +5,7 @@ from temporalio import activity
 
 from posthog.temporal.common.utils import asyncify
 
-from products.tasks.backend.logic.services.sandbox import Sandbox
+from products.tasks.backend.logic.services.sandbox import get_sandbox_class_for_sandbox_id
 from products.tasks.backend.temporal.observability import log_activity_execution
 
 logger = logging.getLogger(__name__)
@@ -24,7 +24,7 @@ def cleanup_sandbox(input: CleanupSandboxInput) -> None:
         sandbox_id=input.sandbox_id,
     ):
         try:
-            sandbox = Sandbox.get_by_id(input.sandbox_id)
+            sandbox = get_sandbox_class_for_sandbox_id(input.sandbox_id).get_by_id(input.sandbox_id)
             sandbox.destroy()
         except Exception:
             # The sandbox has a timeout, and it will eventually terminate if we failed to cleanup
