+++ products/tasks/backend/temporal/create_snapshot/activities/create_snapshot.py
@@ -4,7 +4,7 @@ from temporalio import activity
 
 from posthog.temporal.common.utils import asyncify
 
-from products.tasks.backend.logic.services.sandbox import Sandbox
+from products.tasks.backend.logic.services.sandbox import get_sandbox_class_for_sandbox_id
 from products.tasks.backend.models import SandboxSnapshot
 from products.tasks.backend.temporal.observability import log_activity_execution
 
@@ -27,7 +27,7 @@ def create_snapshot(input: CreateSnapshotInput) -> str:
         sandbox_id=input.sandbox_id,
         **ctx.to_log_context(),
     ):
-        sandbox = Sandbox.get_by_id(input.sandbox_id)
+        sandbox = get_sandbox_class_for_sandbox_id(input.sandbox_id).get_by_id(input.sandbox_id)
 
         snapshot_external_id = sandbox.create_snapshot()
