+++ products/tasks/backend/temporal/execute_sandbox/activities/reap_orphaned_sandbox.py
@@ -22,7 +22,7 @@ from temporalio import activity
 
 from posthog.temporal.common.utils import asyncify
 
-from products.tasks.backend.logic.services.sandbox import Sandbox
+from products.tasks.backend.logic.services.sandbox import get_sandbox_class_for_sandbox_id
 from products.tasks.backend.logic.services.sandbox_usage import (
     close_sandbox_session,
     measure_sandbox_billed_cpu_usage,
@@ -72,7 +72,7 @@ def reap_orphaned_sandbox(input: ReapOrphanedSandboxInput) -> ReapOrphanedSandbo
         billed_cpu_usage_usec = None
         cpu_usage_measured_at = None
         try:
-            sandbox = Sandbox.get_by_id(sandbox_id)
+            sandbox = get_sandbox_class_for_sandbox_id(sandbox_id).get_by_id(sandbox_id)
             cpu_usage_usec, cpu_usage_measured_at = measure_sandbox_cpu_usage(sandbox)
             billed_cpu_usage_usec = measure_sandbox_billed_cpu_usage(sandbox)
             sandbox.destroy()
