+++ products/tasks/backend/temporal/execute_sandbox/activities/tests/test_reap_orphaned_sandbox.py
@@ -11,9 +11,11 @@ from products.tasks.backend.temporal.execute_sandbox.activities.reap_orphaned_sa
 )
 from products.tasks.backend.temporal.execute_sandbox.activities.sandbox_state import SANDBOX_ID_STATE_KEY
 
-# Patch target — `Sandbox` is imported into the reap module's namespace, so
-# patching at the import site is what intercepts the call.
-SANDBOX_IMPORT_PATH = "products.tasks.backend.temporal.execute_sandbox.activities.reap_orphaned_sandbox.Sandbox"
+# Patch target — the reap module resolves the provider class per sandbox id, so
+# patching the resolver at its import site is what intercepts the call.
+SANDBOX_IMPORT_PATH = (
+    "products.tasks.backend.temporal.execute_sandbox.activities.reap_orphaned_sandbox.get_sandbox_class_for_sandbox_id"
+)
 
 
 @pytest.mark.requires_secrets
@@ -29,7 +31,7 @@ class TestReapOrphanedSandbox:
         assert result.reaped_sandbox_id is None
         assert result.destroy_succeeded is True
         # Modal call must not happen when there's nothing to reap.
-        sandbox_cls.get_by_id.assert_not_called()
+        sandbox_cls.return_value.get_by_id.assert_not_called()
 
     def test_returns_none_when_task_run_missing(self, activity_environment):
         with patch(SANDBOX_IMPORT_PATH) as sandbox_cls:
@@ -39,7 +41,7 @@ class TestReapOrphanedSandbox:
             )
 
         assert result.reaped_sandbox_id is None
-        sandbox_cls.get_by_id.assert_not_called()
+        sandbox_cls.return_value.get_by_id.assert_not_called()
 
     @pytest.mark.parametrize("bogus_value", [123, "", None, [], {}])
     def test_treats_non_string_persisted_value_as_no_sandbox(self, activity_environment, test_task_run, bogus_value):
@@ -56,7 +58,7 @@ class TestReapOrphanedSandbox:
             )
 
         assert result.reaped_sandbox_id is None
-        sandbox_cls.get_by_id.assert_not_called()
+        sandbox_cls.return_value.get_by_id.assert_not_called()
 
     def test_destroys_and_clears_when_persisted_id_present(self, activity_environment, test_task_run):
         test_task_run.state = {SANDBOX_ID_STATE_KEY: "sb-orphan", "mode": "background"}
@@ -64,7 +66,7 @@ class TestReapOrphanedSandbox:
 
         with patch(SANDBOX_IMPORT_PATH) as sandbox_cls:
             destroy_mock = Mock()
-            sandbox_cls.get_by_id.return_value = Mock(destroy=destroy_mock)
+            sandbox_cls.return_value.get_by_id.return_value = Mock(destroy=destroy_mock)
 
             result = async_to_sync(activity_environment.run)(
                 reap_orphaned_sandbox,
@@ -73,7 +75,7 @@ class TestReapOrphanedSandbox:
 
         assert result.reaped_sandbox_id == "sb-orphan"
         assert result.destroy_succeeded is True
-        sandbox_cls.get_by_id.assert_called_once_with("sb-orphan")
+        sandbox_cls.return_value.get_by_id.assert_called_once_with("sb-orphan")
         destroy_mock.assert_called_once()
 
         test_task_run.refresh_from_db()
@@ -91,7 +93,7 @@ class TestReapOrphanedSandbox:
                 "products.tasks.backend.temporal.execute_sandbox.activities.reap_orphaned_sandbox.close_sandbox_session"
             ) as close_session,
         ):
-            sandbox = sandbox_cls.get_by_id.return_value
+            sandbox = sandbox_cls.return_value.get_by_id.return_value
             sandbox.read_cpu_usage_usec.return_value = 12_345_678
             sandbox.read_billed_cpu_usage_usec.return_value = 15_000_000
 
@@ -119,7 +121,7 @@ class TestReapOrphanedSandbox:
         test_task_run.save(update_fields=["state"])
 
         with patch(SANDBOX_IMPORT_PATH) as sandbox_cls:
-            sandbox_cls.get_by_id.side_effect = RuntimeError("modal down")
+            sandbox_cls.return_value.get_by_id.side_effect = RuntimeError("modal down")
 
             result = async_to_sync(activity_environment.run)(
                 reap_orphaned_sandbox,
