+++ products/tasks/backend/temporal/process_task/activities/tests/test_refresh_sandbox_credentials.py
@@ -35,8 +35,8 @@ class TestRefreshSandboxCredentialsActivity:
     ):
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.sandbox_credentials.get_sandbox_github_token",
@@ -71,8 +71,8 @@ class TestRefreshSandboxCredentialsActivity:
         test_task.save(update_fields=["state", "updated_at"])
 
         with patch(
-            "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-            return_value=sandbox,
+            "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+            **{"return_value.get_by_id.return_value": sandbox},
         ) as get_sandbox:
             output = async_to_sync(activity_environment.run)(
                 refresh_sandbox_credentials,
@@ -93,8 +93,8 @@ class TestRefreshSandboxCredentialsActivity:
 
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.sandbox_credentials.get_sandbox_github_token",
@@ -118,8 +118,8 @@ class TestRefreshSandboxCredentialsActivity:
                 "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Task"
             ) as mock_task,
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.sandbox_credentials.get_sandbox_github_token",
@@ -144,8 +144,8 @@ class TestRefreshSandboxCredentialsActivity:
     def test_credential_failure_is_non_fatal(self, activity_environment, task_context, test_task, sandbox):
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.sandbox_credentials.get_sandbox_github_token",
@@ -168,8 +168,8 @@ class TestRefreshSandboxCredentialsActivity:
         sandbox.is_running.return_value = False
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.sandbox_credentials.get_sandbox_github_token"
@@ -196,8 +196,8 @@ class TestRefreshSandboxCredentialsActivity:
         # flag the refresh loop stops on, not as an error the loop swallows and retries.
         context = dataclasses.replace(task_context, task_id=str(uuid.uuid4()))
         with patch(
-            "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-            return_value=sandbox,
+            "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+            **{"return_value.get_by_id.return_value": sandbox},
         ):
             output = async_to_sync(activity_environment.run)(
                 refresh_sandbox_credentials,
@@ -214,11 +214,15 @@ class TestRefreshSandboxCredentialsActivity:
         # fire a spurious "task failed" alert after the run's PR is already open).
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                side_effect=SandboxNotFoundError(
-                    "Sandbox sandbox-abc not found",
-                    {"sandbox_id": "sandbox-abc"},
-                    cause=RuntimeError("Deadline Exceeded"),
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                return_value=MagicMock(
+                    get_by_id=MagicMock(
+                        side_effect=SandboxNotFoundError(
+                            "Sandbox sandbox-abc not found",
+                            {"sandbox_id": "sandbox-abc"},
+                            cause=RuntimeError("Deadline Exceeded"),
+                        )
+                    )
                 ),
             ),
             patch(
@@ -252,8 +256,8 @@ class TestRefreshSandboxCredentialsActivity:
         )
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.sandbox_credentials.get_sandbox_github_token",
@@ -282,8 +286,8 @@ class TestRefreshSandboxCredentialsActivity:
         test_task.refresh_from_db()
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch("products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.track_event"),
             patch(
@@ -304,8 +308,8 @@ class TestRefreshSandboxCredentialsActivity:
     def test_excluded_kinds_report_nothing_left(self, activity_environment, task_context, test_task, sandbox):
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.track_event"
@@ -327,11 +331,15 @@ class TestRefreshSandboxCredentialsActivity:
     def test_sandbox_gone_wins_over_excluded_kinds(self, activity_environment, task_context, test_task):
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                side_effect=SandboxNotFoundError(
-                    "Sandbox sandbox-abc not found",
-                    {"sandbox_id": "sandbox-abc"},
-                    cause=RuntimeError("Deadline Exceeded"),
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                return_value=MagicMock(
+                    get_by_id=MagicMock(
+                        side_effect=SandboxNotFoundError(
+                            "Sandbox sandbox-abc not found",
+                            {"sandbox_id": "sandbox-abc"},
+                            cause=RuntimeError("Deadline Exceeded"),
+                        )
+                    )
                 ),
             ),
             patch("products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.track_event"),
@@ -352,8 +360,8 @@ class TestRefreshSandboxCredentialsActivity:
         )
         with (
             patch(
-                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.Sandbox.get_by_id",
-                return_value=sandbox,
+                "products.tasks.backend.temporal.process_task.activities.refresh_sandbox_credentials.get_sandbox_class_for_sandbox_id",
+                **{"return_value.get_by_id.return_value": sandbox},
             ),
             patch(
                 "products.tasks.backend.temporal.process_task.sandbox_credentials.get_sandbox_github_token",
