+++ products/tasks/backend/temporal/process_task/activities/tests/test_start_agent_server.py
@@ -134,8 +134,8 @@ async def test_start_failure_does_not_report_network_enforcement_observation(moc
         network_policy_fingerprint="policy-hash",
     )
     mocker.patch(
-        "products.tasks.backend.temporal.process_task.activities.start_agent_server.Sandbox.get_by_id",
-        return_value=mocker.Mock(),
+        "products.tasks.backend.temporal.process_task.activities.start_agent_server.get_sandbox_class_for_sandbox_id",
+        **{"return_value.get_by_id.return_value": mocker.Mock()},
     )
     mocker.patch(
         "products.tasks.backend.temporal.process_task.activities.start_agent_server._prepare_launch",
@@ -168,8 +168,8 @@ async def test_await_agent_server_ready_relaunches_on_activity_retries(mocker, a
     sandbox.execute.return_value.stdout = ""
     sandbox.read_agent_server_boot_metrics.return_value = (None, {})
     mocker.patch(
-        "products.tasks.backend.temporal.process_task.activities.start_agent_server.Sandbox.get_by_id",
-        return_value=sandbox,
+        "products.tasks.backend.temporal.process_task.activities.start_agent_server.get_sandbox_class_for_sandbox_id",
+        **{"return_value.get_by_id.return_value": sandbox},
     )
     mocker.patch(
         "products.tasks.backend.temporal.process_task.activities.start_agent_server.current_activity_attempt",
@@ -234,8 +234,8 @@ async def test_await_agent_server_ready_records_failed_relaunch(mocker) -> None:
     sandbox = mocker.Mock(id="sandbox-id")
     sandbox.start_agent_server.side_effect = RuntimeError("session did not initialize")
     mocker.patch(
-        "products.tasks.backend.temporal.process_task.activities.start_agent_server.Sandbox.get_by_id",
-        return_value=sandbox,
+        "products.tasks.backend.temporal.process_task.activities.start_agent_server.get_sandbox_class_for_sandbox_id",
+        **{"return_value.get_by_id.return_value": sandbox},
     )
     mocker.patch(
         "products.tasks.backend.temporal.process_task.activities.start_agent_server.current_activity_attempt",
@@ -575,8 +575,8 @@ async def test_start_agent_server_uses_captured_sandbox_event_ingest_flag(mocker
     sandbox.execute.return_value.stderr = ""
     sandbox.read_agent_server_boot_metrics.return_value = (None, {})
     mocker.patch(
-        "products.tasks.backend.temporal.process_task.activities.start_agent_server.Sandbox.get_by_id",
-        return_value=sandbox,
+        "products.tasks.backend.temporal.process_task.activities.start_agent_server.get_sandbox_class_for_sandbox_id",
+        **{"return_value.get_by_id.return_value": sandbox},
     )
     mocker.patch("products.tasks.backend.temporal.process_task.activities.start_agent_server.emit_agent_log")
     task = mocker.Mock(
@@ -649,8 +649,8 @@ async def test_start_agent_server_forwards_imported_and_relayed_mcp_servers(mock
     sandbox.execute.return_value.stderr = ""
     sandbox.read_agent_server_boot_metrics.return_value = (None, {})
     mocker.patch(
-        "products.tasks.backend.temporal.process_task.activities.start_agent_server.Sandbox.get_by_id",
-        return_value=sandbox,
+        "products.tasks.backend.temporal.process_task.activities.start_agent_server.get_sandbox_class_for_sandbox_id",
+        **{"return_value.get_by_id.return_value": sandbox},
     )
     mocker.patch("products.tasks.backend.temporal.process_task.activities.start_agent_server.emit_agent_log")
     mocker.patch(
@@ -708,8 +708,8 @@ async def test_start_agent_server_passes_initial_permission_mode(mocker) -> None
     sandbox.execute.return_value.stderr = ""
     sandbox.read_agent_server_boot_metrics.return_value = (None, {})
     mocker.patch(
-        "products.tasks.backend.temporal.process_task.activities.start_agent_server.Sandbox.get_by_id",
-        return_value=sandbox,
+        "products.tasks.backend.temporal.process_task.activities.start_agent_server.get_sandbox_class_for_sandbox_id",
+        **{"return_value.get_by_id.return_value": sandbox},
     )
     mocker.patch("products.tasks.backend.temporal.process_task.activities.start_agent_server.emit_agent_log")
     mocker.patch(
